+++ src/main/java/com/example/rinhabackend/repository/ExtratoRepository.java
@@ -16,7 +16,7 @@ import java.util.Objects;
 
 public class ExtratoRepository {
 
-    public ExtratoResponse findAllById(Long idCliente) {
+    public ExtratoResponse findAllById(int idCliente) {
         PreparedStatement prepareStatement;
 
         ExtratoResponse extratoResponse = new ExtratoResponse();
@@ -27,7 +27,7 @@ public class ExtratoRepository {
 
             prepareStatement = connection.prepareStatement("SELECT c.saldo, c.limite, t.valor, t.tipo, t.descricao, t.realizada_em from cliente c left join transacao t " +
                     "ON t.id_cliente = c.id WHERE c.id = ? order by t.realizada_em DESC limit 10");
-            prepareStatement.setLong(1, idCliente);
+            prepareStatement.setInt(1, idCliente);
 
             boolean extratoResponseNaoPreenchido = true;
